refactor: Support new Verifications 'Report' and 'Status' dedicated tags from OAS
