+++ client/src/main/com/sinch/sdk/domains/verification/adapters/VerificationsReportService.java
@@ -1,0 +1,81 @@
+package com.sinch.sdk.domains.verification.adapters;
+
+import com.sinch.sdk.core.http.AuthManager;
+import com.sinch.sdk.core.http.HttpClient;
+import com.sinch.sdk.core.http.HttpMapper;
+import com.sinch.sdk.domains.verification.adapters.converters.VerificationsDtoConverter;
+import com.sinch.sdk.domains.verification.api.v1.internal.VerificationsReportApi;
+import com.sinch.sdk.domains.verification.models.NumberIdentity;
+import com.sinch.sdk.domains.verification.models.VerificationId;
+import com.sinch.sdk.domains.verification.models.VerificationReport;
+import com.sinch.sdk.domains.verification.models.VerificationReportCallout;
+import com.sinch.sdk.domains.verification.models.VerificationReportFlashCall;
+import com.sinch.sdk.domains.verification.models.VerificationReportSMS;
+import com.sinch.sdk.domains.verification.models.requests.VerificationReportCalloutRequestParameters;
+import com.sinch.sdk.domains.verification.models.requests.VerificationReportFlashCallRequestParameters;
+import com.sinch.sdk.domains.verification.models.requests.VerificationReportRequestParameters;
+import com.sinch.sdk.domains.verification.models.requests.VerificationReportSMSRequestParameters;
+import com.sinch.sdk.models.VerificationContext;
+import java.util.Map;
+
+public class VerificationsReportService {
+
+  private final VerificationsReportApi api;
+
+  public VerificationsReportService(
+      VerificationContext context, HttpClient httpClient, Map<String, AuthManager> authManagers) {
+    this.api =
+        new VerificationsReportApi(
+            httpClient, context.getVerificationServer(), authManagers, new HttpMapper());
+  }
+
+  protected VerificationsReportApi getApi() {
+    return this.api;
+  }
+
+  public VerificationReportSMS reportSmsByIdentity(
+      NumberIdentity identity, VerificationReportSMSRequestParameters parameters) {
+    return (VerificationReportSMS) report(identity, parameters);
+  }
+
+  public VerificationReportFlashCall reportFlashCallByIdentity(
+      NumberIdentity identity, VerificationReportFlashCallRequestParameters parameters) {
+    return (VerificationReportFlashCall) report(identity, parameters);
+  }
+
+  public VerificationReportCallout reportCalloutByIdentity(
+      NumberIdentity identity, VerificationReportCalloutRequestParameters parameters) {
+    return (VerificationReportCallout) report(identity, parameters);
+  }
+
+  private VerificationReport report(
+      NumberIdentity identity, VerificationReportRequestParameters parameters) {
+
+    return VerificationsDtoConverter.convert(
+        getApi()
+            .reportVerificationByIdentity(
+                identity.getEndpoint(), VerificationsDtoConverter.convert(parameters)));
+  }
+
+  public VerificationReportSMS reportSmsById(
+      VerificationId id, VerificationReportSMSRequestParameters parameters) {
+    return (VerificationReportSMS) report(id, parameters);
+  }
+
+  public VerificationReportFlashCall reportFlashCallById(
+      VerificationId id, VerificationReportFlashCallRequestParameters parameters) {
+    return (VerificationReportFlashCall) report(id, parameters);
+  }
+
+  public VerificationReportCallout reportCalloutById(
+      VerificationId id, VerificationReportCalloutRequestParameters parameters) {
+    return (VerificationReportCallout) report(id, parameters);
+  }
+
+  private VerificationReport report(
+      VerificationId id, VerificationReportRequestParameters parameters) {
+
+    return VerificationsDtoConverter.convert(
+        getApi().reportVerificationById(id.getId(), VerificationsDtoConverter.convert(parameters)));
+  }
+}
