+++ client/src/main/com/sinch/sdk/domains/verification/adapters/VerificationsService.java
@@ -2,25 +2,18 @@ package com.sinch.sdk.domains.verification.adapters;
 
 import com.sinch.sdk.core.http.AuthManager;
 import com.sinch.sdk.core.http.HttpClient;
-import com.sinch.sdk.core.http.HttpMapper;
-import com.sinch.sdk.domains.verification.adapters.converters.VerificationsDtoConverter;
-import com.sinch.sdk.domains.verification.api.v1.internal.VerificationsApi;
 import com.sinch.sdk.domains.verification.models.NumberIdentity;
 import com.sinch.sdk.domains.verification.models.VerificationId;
-import com.sinch.sdk.domains.verification.models.VerificationReport;
 import com.sinch.sdk.domains.verification.models.VerificationReportCallout;
 import com.sinch.sdk.domains.verification.models.VerificationReportFlashCall;
 import com.sinch.sdk.domains.verification.models.VerificationReportSMS;
 import com.sinch.sdk.domains.verification.models.requests.StartVerificationCalloutRequestParameters;
 import com.sinch.sdk.domains.verification.models.requests.StartVerificationFlashCallRequestParameters;
-import com.sinch.sdk.domains.verification.models.requests.StartVerificationRequestParameters;
 import com.sinch.sdk.domains.verification.models.requests.StartVerificationSMSRequestParameters;
 import com.sinch.sdk.domains.verification.models.requests.StartVerificationSeamlessRequestParameters;
 import com.sinch.sdk.domains.verification.models.requests.VerificationReportCalloutRequestParameters;
 import com.sinch.sdk.domains.verification.models.requests.VerificationReportFlashCallRequestParameters;
-import com.sinch.sdk.domains.verification.models.requests.VerificationReportRequestParameters;
 import com.sinch.sdk.domains.verification.models.requests.VerificationReportSMSRequestParameters;
-import com.sinch.sdk.domains.verification.models.response.StartVerificationResponse;
 import com.sinch.sdk.domains.verification.models.response.StartVerificationResponseCallout;
 import com.sinch.sdk.domains.verification.models.response.StartVerificationResponseFlashCall;
 import com.sinch.sdk.domains.verification.models.response.StartVerificationResponseSMS;
@@ -31,97 +24,61 @@ import java.util.Map;
 public class VerificationsService
     implements com.sinch.sdk.domains.verification.VerificationsService {
 
-  private final VerificationsApi api;
+  private final VerificationsReportService reportService;
+  private final VerificationsStartService startService;
 
   public VerificationsService(
       VerificationContext context, HttpClient httpClient, Map<String, AuthManager> authManagers) {
-    this.api =
-        new VerificationsApi(
-            httpClient, context.getVerificationServer(), authManagers, new HttpMapper());
-  }
-
-  protected VerificationsApi getApi() {
-    return this.api;
+    this.reportService = new VerificationsReportService(context, httpClient, authManagers);
+    this.startService = new VerificationsStartService(context, httpClient, authManagers);
   }
 
   public StartVerificationResponseSMS startSms(StartVerificationSMSRequestParameters parameters) {
-    String acceptLanguage = null;
-    if (parameters.getOptions().isPresent()
-        && parameters.getOptions().get().getAcceptLanguage().isPresent()) {
-      acceptLanguage = parameters.getOptions().get().getAcceptLanguage().get();
-    }
-    return (StartVerificationResponseSMS) start(parameters, acceptLanguage);
+    return startService.startSms(parameters);
   }
 
   public StartVerificationResponseFlashCall startFlashCall(
       StartVerificationFlashCallRequestParameters parameters) {
-    return (StartVerificationResponseFlashCall) start(parameters);
+    return startService.startFlashCall(parameters);
   }
 
   public StartVerificationResponseCallout startCallout(
       StartVerificationCalloutRequestParameters parameters) {
-    return (StartVerificationResponseCallout) start(parameters);
+    return startService.startCallout(parameters);
   }
 
   public StartVerificationResponseSeamless startSeamless(
       StartVerificationSeamlessRequestParameters parameters) {
-    return (StartVerificationResponseSeamless) start(parameters);
-  }
-
-  private StartVerificationResponse start(StartVerificationRequestParameters parameters) {
-    return VerificationsDtoConverter.convert(
-        getApi().startVerification(VerificationsDtoConverter.convert(parameters), null));
-  }
-
-  private StartVerificationResponse start(
-      StartVerificationRequestParameters parameters, String acceptLanguage) {
-    return VerificationsDtoConverter.convert(
-        getApi().startVerification(VerificationsDtoConverter.convert(parameters), acceptLanguage));
+    return startService.startSeamless(parameters);
   }
 
   public VerificationReportSMS reportSmsByIdentity(
       NumberIdentity identity, VerificationReportSMSRequestParameters parameters) {
-    return (VerificationReportSMS) report(identity, parameters);
+    return reportService.reportSmsByIdentity(identity, parameters);
   }
 
   public VerificationReportFlashCall reportFlashCallByIdentity(
       NumberIdentity identity, VerificationReportFlashCallRequestParameters parameters) {
-    return (VerificationReportFlashCall) report(identity, parameters);
+    return reportService.reportFlashCallByIdentity(identity, parameters);
   }
 
   public VerificationReportCallout reportCalloutByIdentity(
       NumberIdentity identity, VerificationReportCalloutRequestParameters parameters) {
-    return (VerificationReportCallout) report(identity, parameters);
-  }
-
-  private VerificationReport report(
-      NumberIdentity identity, VerificationReportRequestParameters parameters) {
-
-    return VerificationsDtoConverter.convert(
-        getApi()
-            .reportVerificationByIdentity(
-                identity.getEndpoint(), VerificationsDtoConverter.convert(parameters)));
+    return reportService.reportCalloutByIdentity(identity, parameters);
   }
 
   public VerificationReportSMS reportSmsById(
       VerificationId id, VerificationReportSMSRequestParameters parameters) {
-    return (VerificationReportSMS) report(id, parameters);
+    return reportService.reportSmsById(id, parameters);
   }
 
   public VerificationReportFlashCall reportFlashCallById(
       VerificationId id, VerificationReportFlashCallRequestParameters parameters) {
-    return (VerificationReportFlashCall) report(id, parameters);
+    return reportService.reportFlashCallById(id, parameters);
   }
 
   public VerificationReportCallout reportCalloutById(
       VerificationId id, VerificationReportCalloutRequestParameters parameters) {
-    return (VerificationReportCallout) report(id, parameters);
-  }
-
-  private VerificationReport report(
-      VerificationId id, VerificationReportRequestParameters parameters) {
-
-    return VerificationsDtoConverter.convert(
-        getApi().reportVerificationById(id.getId(), VerificationsDtoConverter.convert(parameters)));
+    return reportService.reportCalloutById(id, parameters);
   }
 }
