+++ client/src/main/com/sinch/sdk/domains/verification/adapters/VerificationsStartService.java
@@ -1,0 +1,70 @@
+package com.sinch.sdk.domains.verification.adapters;
+
+import com.sinch.sdk.core.http.AuthManager;
+import com.sinch.sdk.core.http.HttpClient;
+import com.sinch.sdk.core.http.HttpMapper;
+import com.sinch.sdk.domains.verification.adapters.converters.VerificationsDtoConverter;
+import com.sinch.sdk.domains.verification.api.v1.internal.VerificationsStartApi;
+import com.sinch.sdk.domains.verification.models.requests.StartVerificationCalloutRequestParameters;
+import com.sinch.sdk.domains.verification.models.requests.StartVerificationFlashCallRequestParameters;
+import com.sinch.sdk.domains.verification.models.requests.StartVerificationRequestParameters;
+import com.sinch.sdk.domains.verification.models.requests.StartVerificationSMSRequestParameters;
+import com.sinch.sdk.domains.verification.models.requests.StartVerificationSeamlessRequestParameters;
+import com.sinch.sdk.domains.verification.models.response.StartVerificationResponse;
+import com.sinch.sdk.domains.verification.models.response.StartVerificationResponseCallout;
+import com.sinch.sdk.domains.verification.models.response.StartVerificationResponseFlashCall;
+import com.sinch.sdk.domains.verification.models.response.StartVerificationResponseSMS;
+import com.sinch.sdk.domains.verification.models.response.StartVerificationResponseSeamless;
+import com.sinch.sdk.models.VerificationContext;
+import java.util.Map;
+
+public class VerificationsStartService {
+
+  private final VerificationsStartApi api;
+
+  public VerificationsStartService(
+      VerificationContext context, HttpClient httpClient, Map<String, AuthManager> authManagers) {
+    this.api =
+        new VerificationsStartApi(
+            httpClient, context.getVerificationServer(), authManagers, new HttpMapper());
+  }
+
+  protected VerificationsStartApi getApi() {
+    return this.api;
+  }
+
+  public StartVerificationResponseSMS startSms(StartVerificationSMSRequestParameters parameters) {
+    String acceptLanguage = null;
+    if (parameters.getOptions().isPresent()
+        && parameters.getOptions().get().getAcceptLanguage().isPresent()) {
+      acceptLanguage = parameters.getOptions().get().getAcceptLanguage().get();
+    }
+    return (StartVerificationResponseSMS) start(parameters, acceptLanguage);
+  }
+
+  public StartVerificationResponseFlashCall startFlashCall(
+      StartVerificationFlashCallRequestParameters parameters) {
+    return (StartVerificationResponseFlashCall) start(parameters);
+  }
+
+  public StartVerificationResponseCallout startCallout(
+      StartVerificationCalloutRequestParameters parameters) {
+    return (StartVerificationResponseCallout) start(parameters);
+  }
+
+  public StartVerificationResponseSeamless startSeamless(
+      StartVerificationSeamlessRequestParameters parameters) {
+    return (StartVerificationResponseSeamless) start(parameters);
+  }
+
+  private StartVerificationResponse start(StartVerificationRequestParameters parameters) {
+    return VerificationsDtoConverter.convert(
+        getApi().startVerification(VerificationsDtoConverter.convert(parameters), null));
+  }
+
+  private StartVerificationResponse start(
+      StartVerificationRequestParameters parameters, String acceptLanguage) {
+    return VerificationsDtoConverter.convert(
+        getApi().startVerification(VerificationsDtoConverter.convert(parameters), acceptLanguage));
+  }
+}
