+++ client/src/test/java/com/sinch/sdk/domains/verification/adapters/VerificationsReportServiceTest.java
@@ -12,20 +12,14 @@ import com.sinch.sdk.core.exceptions.ApiException;
 import com.sinch.sdk.core.http.AuthManager;
 import com.sinch.sdk.core.http.HttpClient;
 import com.sinch.sdk.domains.verification.adapters.converters.VerificationsDtoConverterTest;
-import com.sinch.sdk.domains.verification.api.v1.internal.VerificationsApi;
+import com.sinch.sdk.domains.verification.api.v1.internal.VerificationsReportApi;
 import com.sinch.sdk.domains.verification.models.NumberIdentity;
 import com.sinch.sdk.domains.verification.models.VerificationId;
 import com.sinch.sdk.domains.verification.models.VerificationReportCallout;
 import com.sinch.sdk.domains.verification.models.VerificationReportFlashCall;
 import com.sinch.sdk.domains.verification.models.VerificationReportSMS;
 import com.sinch.sdk.domains.verification.models.dto.v1.report.VerificationReportResponseDtoTest;
-import com.sinch.sdk.domains.verification.models.dto.v1.start.StartVerificationResponseTest;
-import com.sinch.sdk.domains.verification.models.response.StartVerificationResponseCallout;
-import com.sinch.sdk.domains.verification.models.response.StartVerificationResponseFlashCall;
-import com.sinch.sdk.domains.verification.models.response.StartVerificationResponseSMS;
-import com.sinch.sdk.domains.verification.models.response.StartVerificationResponseSeamless;
 import com.sinch.sdk.domains.verification.models.v1.report.request.internal.VerificationReportRequestParametersInternal;
-import com.sinch.sdk.domains.verification.models.v1.start.request.StartVerificationRequestParameters;
 import com.sinch.sdk.models.VerificationContext;
 import java.util.Map;
 import org.assertj.core.api.Assertions;
@@ -34,19 +28,7 @@ import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 
 @TestWithResources
-public class VerificationsServiceTest extends BaseTest {
-
-  @GivenJsonResource("/domains/verification/v1/start/StartVerificationCalloutRequestDto.json")
-  public StartVerificationRequestParameters startVerificationCalloutRequestDto;
-
-  @GivenJsonResource("/domains/verification/v1/start/StartVerificationFlashCallRequestDto.json")
-  public StartVerificationRequestParameters startVerificationFlashCallRequestDto;
-
-  @GivenJsonResource("/domains/verification/v1/start/StartVerificationSeamlessRequestDto.json")
-  public StartVerificationRequestParameters startVerificationSeamlessRequestDto;
-
-  @GivenJsonResource("/domains/verification/v1/start/StartVerificationSMSRequestDto.json")
-  public StartVerificationRequestParameters startVerificationSMSRequestDto;
+public class VerificationsReportServiceTest extends BaseTest {
 
   @GivenJsonResource("/domains/verification/v1/report/VerificationReportCalloutRequestDto.json")
   public VerificationReportRequestParametersInternal verificationReportCalloutRequestDto;
@@ -57,73 +39,17 @@ public class VerificationsServiceTest extends BaseTest {
   @GivenJsonResource("/domains/verification/v1/report/VerificationReportSMSRequestDto.json")
   public VerificationReportRequestParametersInternal verificationReportSMSRequestDto;
 
-  @Mock VerificationsApi api;
+  @Mock VerificationsReportApi api;
   @Mock VerificationContext context;
   @Mock HttpClient httpClient;
   @Mock Map<String, AuthManager> authManagers;
 
-  VerificationsService service;
+  VerificationsReportService service;
 
   @BeforeEach
   public void initMocks() {
-    service = spy(new VerificationsService(context, httpClient, authManagers));
+    service = spy(new VerificationsReportService(context, httpClient, authManagers));
     doReturn(api).when(service).getApi();
-  }
-
-  @Test
-  void startSms() throws ApiException {
-
-    when(api.startVerification(eq(startVerificationSMSRequestDto), eq(null)))
-        .thenReturn(StartVerificationResponseTest.expectedStartVerificationSMSDto);
-
-    StartVerificationResponseSMS response =
-        service.startSms(VerificationsDtoConverterTest.startVerificationSMSRequest);
-
-    Assertions.assertThat(response)
-        .usingRecursiveComparison()
-        .isEqualTo(VerificationsDtoConverterTest.expectedStartVerificationSMSResponse);
-  }
-
-  @Test
-  void startFlashCall() throws ApiException {
-
-    when(api.startVerification(eq(startVerificationFlashCallRequestDto), eq(null)))
-        .thenReturn(StartVerificationResponseTest.expectedStartVerificationFlashCallDto);
-
-    StartVerificationResponseFlashCall response =
-        service.startFlashCall(VerificationsDtoConverterTest.startVerificationFlashCallRequest);
-
-    Assertions.assertThat(response)
-        .usingRecursiveComparison()
-        .isEqualTo(VerificationsDtoConverterTest.expectedStartVerificationFlashCallResponse);
-  }
-
-  @Test
-  void startCallout() throws ApiException {
-
-    when(api.startVerification(eq(startVerificationCalloutRequestDto), eq(null)))
-        .thenReturn(StartVerificationResponseTest.expectedStartVerificationCalloutDto);
-
-    StartVerificationResponseCallout response =
-        service.startCallout(VerificationsDtoConverterTest.startVerificationCalloutRequest);
-
-    Assertions.assertThat(response)
-        .usingRecursiveComparison()
-        .isEqualTo(VerificationsDtoConverterTest.expectedStartVerificationCalloutResponse);
-  }
-
-  @Test
-  void startSeamless() throws ApiException {
-
-    when(api.startVerification(eq(startVerificationSeamlessRequestDto), eq(null)))
-        .thenReturn(StartVerificationResponseTest.expectedStartVerificationSeamlessDto);
-
-    StartVerificationResponseSeamless response =
-        service.startSeamless(VerificationsDtoConverterTest.startVerificationSeamlessRequest);
-
-    Assertions.assertThat(response)
-        .usingRecursiveComparison()
-        .isEqualTo(VerificationsDtoConverterTest.expectedStartVerificationSeamlessResponse);
   }
 
   @Test
